feat(reference): fold Brands + Pen/Driver Compat into Reference sections
Lands the work from #268 and #269 onto main. Both were merged into their
stacked base branches rather than main (GitHub didn't retarget since the
intermediate branches still existed), so the section consolidation never
reached main. This is the net change as a single clean commit.

- Pen + Driver Compatibility become Reference sections under a
  "Compatibility" category; the /pen-compat and /wacom-driver-compat routes
  are deleted; new DriverCompatSection.svelte holds the update.xml→tablet
  join (#268).
- Brands becomes a Reference section under an "Entities" category; the
  /brands list route is deleted (the /brands/[entityId] redirect stays),
  with brand names still linking to /entity/<brand> (#269).
- dataSubNavTabs() trimmed to Reference / Data Dictionary / API Explorer /
  Data Quality; Nav Data altActive updated; data-dictionary uses
  dataSubNavTabs(); smoke.spec.ts drops the dead route entries and repoints
  the brands list->detail test at /reference#brands.

check 0/0; tree is byte-identical to the verified #269 tip (558 unit, 26
e2e, preview-confirmed there).
