+++ src/lib/nav/subnav-tabs.ts
@@ -21,12 +21,9 @@ export function tabletSubNavTabs(opts: { flaggedCount?: number } = {}): SubNavTa
 export function dataSubNavTabs(): SubNavTab[] {
 	return [
 		{ href: '/reference', label: 'Reference' },
-		{ href: '/brands', label: 'Brands' },
 		{ href: '/data-dictionary', label: 'Data Dictionary' },
 		{ href: '/api-explorer', label: 'API Explorer' },
 		{ href: '/data-quality', label: 'Data Quality' },
-		{ href: '/pen-compat', label: 'Pen Compat' },
-		{ href: '/wacom-driver-compat', label: 'Driver Compat' },
 	];
 }
