--- src/routes/brands/+page.ts
@@ -1,5 +1,0 @@
-export async function load({ parent }) {
-	const { ds } = await parent();
-	const brands = await ds.Brands.toArray();
-	return { brands };
-}
